GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - master (refs/gitbutler/master)
   branch head: 803c5609f09886122fc03f81c3298f24261dc859
   - .github/workflows/update.yml
   - packages/common/sysdo/do.py
   - nvfetcher.toml

Your previous branch was: refs/heads/master

The sha for that commit was: 803c5609f09886122fc03f81c3298f24261dc859

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch